Create Server.c
I am adding the Semaphore. Rn I need to write the main code to only go through the semaphore after it has connected, write the section for sys_Call 2 to unlock, ack, and move on to the next person if they exist. Along with that, I need to write the write()'s and maybe a read or 2. After this I will move to the client and design it from there. I got my 2 methods for the add/remove of the queue from CodeVault. smh, why doesn't c have a double linked list library.